# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.2.0
  description: kri8 API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /ideas:
    get:
      operationId: listIdeas
      tags: [ideas]
      summary: List all ideas for the current user
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: is_used
          in: query
          schema:
            type: boolean
        - name: parent_id
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: List of ideas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Idea"
    post:
      operationId: createIdea
      tags: [ideas]
      summary: Create a new idea
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IdeaInput"
      responses:
        "201":
          description: Created idea
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Idea"
...
  /ideas/recent:
    get:
      operationId: getRecentIdeas
      tags: [ideas]
      summary: Get recently created ideas
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Recent ideas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Idea"
  /ideas/calendar:
    get:
      operationId: getCalendarIdeas
      tags: [ideas]
      summary: Get ideas for a specific month (calendar view)
      parameters:
        - name: month
          in: query
          required: true
          schema:
            type: string
          description: "Month in YYYY-MM format"
      responses:
        "200":
          description: Ideas in the month
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Idea"
  /ideas/{id}:
    get:
      operationId: getIdea
      tags: [ideas]
      summary: Get a single idea with its branches
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Idea detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IdeaDetail"
        "404":
          description: Not found
    patch:
      operationId: updateIdea
      tags: [ideas]
      summary: Update an idea
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IdeaUpdate"
      responses:
        "200":
          description: Updated idea
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Idea"
    delete:
      operationId: deleteIdea
      tags: [ideas]
      summary: Delete an idea and all its branches
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /ideas/{id}/mark-used:
    post:
      operationId: markIdeaUsed
      tags: [ideas]
      summary: Mark an idea as used
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MarkUsedInput"
      responses:
        "200":
          description: Updated idea
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Idea"
  /ideas/{id}/branches:
    get:
      operationId: listBranches
      tags: [ideas]
      summary: List all branches of an idea
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of branches
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Idea"
    post:
      operationId: createBranch
      tags: [ideas]
      summary: Create a branch of an idea
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BranchInput"
      responses:
        "201":
          description: Created branch
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Idea"
...
components:
  schemas:
...
    Idea:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        title:
          type: string
        insight:
          type: ["string", "null"]
        origin:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        videoEditingNotes:
          type: ["string", "null"]
        createdDate:
          type: string
        usedDate:
          type: ["string", "null"]
        customDate:
          type: ["string", "null"]
        isUsed:
          type: boolean
        parentIdeaId:
          type: ["integer", "null"]
        branchCount:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, userId, title, isUsed, branchCount, createdDate, createdAt, updatedAt]
    IdeaDetail:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        title:
          type: string
        insight:
          type: ["string", "null"]
        origin:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        videoEditingNotes:
          type: ["string", "null"]
        createdDate:
          type: string
        usedDate:
          type: ["string", "null"]
        customDate:
          type: ["string", "null"]
        isUsed:
          type: boolean
        parentIdeaId:
          type: ["integer", "null"]
        branchCount:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
        branches:
          type: array
          items:
            $ref: "#/components/schemas/Idea"
      required: [id, userId, title, isUsed, branchCount, createdDate, createdAt, updatedAt, branches]
    IdeaInput:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        insight:
          type: string
        origin:
          type: string
        notes:
          type: string
        videoEditingNotes:
          type: string
        customDate:
          type: string
      required: [title]
    IdeaUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        insight:
          type: string
        origin:
          type: string
        notes:
          type: string
        videoEditingNotes:
          type: string
        customDate:
          type: string
        usedDate:
          type: string
        isUsed:
          type: boolean
        themePreference:
          type: string
    BranchInput:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        insight:
          type: string
        notes:
          type: string
        videoEditingNotes:
          type: string
      required: [title]
    MarkUsedInput:
      type: object
      properties:
        usedDate:
          type: string
...