# lib/api-zod/api-zod/src/generated/types/branchInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * kri8 API specification
 * OpenAPI spec version: 0.2.0
 */

export interface BranchInput {
  /** @minLength 1 */
  title: string;
  insight?: string;
  notes?: string;
  videoEditingNotes?: string;
}